pyproject.toml
[tool.poetry]
name = "crewai-testing"
version = "0.1.0"
description = "Autonomous Agent-Based Testing Framework with Mutation Testing and Quality Assessment"
authors = ["AdamG-74 <gillespieads@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<=3.13"
crewai = "^0.28.0"
langchain = "^0.1.0"
litellm = "^1.0.0"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
coverage = "^7.3.0"
mutmut = "^3.1.0"
networkx = "^3.2.0"
graphviz = "^0.20.0"
radon = "^6.0.0"
pyyaml = "^6.0.0"
markdown = "^3.5.0"
pydantic = "^2.5.0"
rich = "^13.7.0"
click = "^8.1.0"
jinja2 = "^3.1.0"
weasyprint = "^60.0.0"
python-dotenv = "^1.0.0"
ipykernel = "^6.29.5"
jupyter = "^1.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.0.0"
flake8 = "^6.0.0"
mypy = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src --cov-report=html --cov-report=term-missing"

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/test_*"]

[tool.mutmut]
paths_to_mutate = ["src"]
backup = false
runner = "python -m pytest"
dict_synonyms = ["Struct", "NamedStruct"]
